+++ core/9_Misc/enable_console.lua
@@ -1,5 +1,0 @@
--- Enable console
-
-if gm.instance_number(gm.constants.oConsole) == 0 then
-	gm.instance_create_depth(0, 0, -100000001, gm.constants.oConsole)
-end